Add the concept of an "action" to the app
A queue of actions is built up during an event and executed in order at the end of the event. This avoids having to walk the whole tree and makes it easie to manage common functions.

With event types and data as strings, it will also make it possible to create a simple interceptor for instrumentation.